Updates
Updated  code to with the latest build of Xcode, 7.1.1 with a
Deployment Target of iOS 8.0. This was also tested against Wordpress
4.3.1. Please note that you will need to enter your URL, Username, and
Password to have this demo project work. I recommend creating a RPC
account so you do not need to store your main credentials in this app.
Lastly, XMLRPC is enabled by default in Wordpress so if you do not see
the option to enable, its because you using a newer version of
Wordpress and its already enabled.